+++ SplunkAppForWazuh/appserver/static/js/config/routes/settings-states.js
@@ -24,6 +24,23 @@ define(['../module'], function(module) {
             '/static/app/SplunkAppForWazuh/js/controllers/settings/about/about.html',
           onEnter: $navigationService => {
             $navigationService.storeRoute('settings.about')
+          },
+          controller: 'aboutCtrl',
+          resolve: {
+            appInfo: [
+              '$requestService',
+              async ($requestService, $state) => {
+                try {
+                  const result = await $requestService.httpReq(
+                    'GET',
+                    '/manager/app_info'
+                  )
+                  return result.data
+                } catch (error) {
+                  $state.go('settings.api')
+                }
+              }
+            ]
           }
         })
         // settings -> api