+++ SplunkAppForWazuh/appserver/static/js/controllers/agents/ciscat/agentsCiscatCtrl.js
@@ -4,8 +4,9 @@ define([
   '../../../services/visualizations/chart/linear-chart',
   '../../../services/visualizations/table/table',
   '../../../services/visualizations/inputs/time-picker',
-  '../../../services/visualizations/search/search-handler'
-], function(app, ColumnChart, LinearChart, Table, TimePicker, SearchHandler) {
+  '../../../services/visualizations/search/search-handler',
+  '../../../services/rawTableData/rawTableDataService'
+], function(app, ColumnChart, LinearChart, Table, TimePicker, SearchHandler, rawTableDataService) {
   'use strict'
 
   class AgentsCiscat {
@@ -16,10 +17,13 @@ define([
      * @param {*} $state 
      * @param {*} $currentDataService 
      * @param {Object} agent 
+     * @param {*} $reportingService
      */
-    constructor($urlTokenModel, $scope, $state, $currentDataService, agent) {
+    constructor($urlTokenModel, $scope, $state, $currentDataService, agent, $reportingService) {
       this.state = $state
       this.currentDataService = $currentDataService
+      this.reportingService = $reportingService      
+      this.tableResults = {}
       if (!this.currentDataService.getCurrentAgent()) {
         this.state.go('overview')
       }
@@ -141,23 +145,89 @@ define([
           `${
             this.filters
           } sourcetype=wazuh rule.groups="ciscat" | top data.cis.group`,
-          'topCiscatGroups'
+          'topCiscatGroups',
+          this.scope
         ),
         new LinearChart(
           'scanResultEvolution',
           `${
             this.filters
           } sourcetype=wazuh rule.groups="ciscat" | timechart count by data.cis.result usenull=f`,
-          'scanResultEvolution'
+          'scanResultEvolution',
+          this.scope
         ),
         new Table(
           'alertsSummary',
           `${
             this.filters
           } sourcetype=wazuh rule.groups="ciscat" | stats count sparkline by data.cis.rule_title, data.cis.remediation,data.cis.group | sort count desc | rename "data.cis.rule_title" as "Title",  "data.cis.remediation" as "Remediation",  "data.cis.group" as "Group" `,
-          'alertsSummary'
+          'alertsSummary',
+          this.scope
         )
       ]
+
+      this.alertsSummaryTable = new rawTableDataService(
+        'alertsSummaryTable',
+        `${
+          this.filters
+        } sourcetype=wazuh rule.groups="ciscat" | stats count sparkline by data.cis.rule_title, data.cis.remediation,data.cis.group | sort count desc | rename "data.cis.rule_title" as "Title",  "data.cis.remediation" as "Remediation",  "data.cis.group" as "Group" `,
+        'alertsSummaryTableToken',
+        '$result$',
+        this.scope
+      )
+      this.vizz.push(this.alertsSummaryTable)
+
+      this.alertsSummaryTable.getSearch().on('result', (result) => {
+        this.tableResults['Alerts Summary'] = result
+      })
+
+      // Set agent info
+      try {
+        this.agentReportData = {
+          ID: this.agent.data.data.id,
+          Name: this.agent.data.data.name,
+          IP: this.agent.data.data.ip,
+          Version: this.agent.data.data.version,
+          Manager: this.agent.data.data.manager,
+          OS: this.agent.data.data.os.name,
+          dateAdd: this.agent.data.data.dateAdd,
+          lastKeepAlive: this.agent.data.data.lastKeepAlive,
+          group: this.agent.data.data.group.toString()
+        }
+      } catch (error) {
+        this.agentReportData = false
+      }
+
+      /**
+       * Generates report
+       */
+      this.scope.startVis2Png = () =>
+      this.reportingService.startVis2Png('agents-ciscat', 'CIS-CAT', this.filters, [
+        'topCiscatGroups',
+        'scanResultEvolution',
+        'alertsSummary'
+      ],
+      this.reportMetrics,
+      this.tableResults,
+      this.agentReportData
+      )
+
+      this.scope.$on('loadingReporting', (event, data) => {
+        this.scope.loadingReporting = data.status
+      })
+
+      this.scope.$on("checkReportingStatus", () => {
+        this.vizzReady = !this.vizz.filter( v => {
+          return v.finish === false
+        }).length
+        if (this.vizzReady) { 
+          this.scope.loadingVizz = false
+          this.setReportMetrics()
+        } else { 
+          this.scope.loadingVizz = true
+        }
+        if (!this.scope.$$phase) this.scope.$digest()
+      })
     }
 
     /**
@@ -202,6 +272,22 @@ define([
      */
     getAgentStatusClass(agentStatus) {
       agentStatus === 'Active' ? 'teal' : 'red'
+    }
+
+    /**
+     * Set report metrics
+     */
+    setReportMetrics() {
+      this.reportMetrics = {
+        'Last not checked': this.scope.lastNotChecked,
+        'Last pass': this.scope.lastPass,
+        'Last scan score': this.scope.lastScanScore,
+        'Last scan date': this.scope.lastScanDate,
+        'Last errores': this.scope.lastErrors,
+        'Last fails': this.scope.lastFails,
+        'Last unknown': this.scope.lastUnknown,
+        'Last scan benchmark': this.scope.lastScanBenchmark
+      }
     }
 
     /**