+++ SplunkAppForWazuh/appserver/static/js/controllers/agents/gdpr/agentsGdprCtrl.js
@@ -4,8 +4,9 @@ define([
   '../../../services/visualizations/chart/pie-chart',
   '../../../services/visualizations/table/table',
   '../../../services/visualizations/inputs/time-picker',
-  '../../../services/visualizations/inputs/dropdown-input'
-], function(app, ColumnChart, PieChart, Table, TimePicker, Dropdown) {
+  '../../../services/visualizations/inputs/dropdown-input',
+  '../../../services/rawTableData/rawTableDataService'
+], function(app, ColumnChart, PieChart, Table, TimePicker, Dropdown, rawTableDataService) {
   'use strict'
 
   class AgentsGdpr {
@@ -16,12 +17,15 @@ define([
      * @param {Object} $currentDataService
      * @param {Object} $state
      * @param {Object} agent
+     * @param {*} $reportingService 
      */
 
-    constructor($urlTokenModel, $currentDataService, $scope, $state, agent) {
+    constructor($urlTokenModel, $currentDataService, $scope, $state, agent, $reportingService) {
       this.scope = $scope
       this.state = $state
       this.currentDataService = $currentDataService
+      this.reportingService = $reportingService
+      this.tableResults = {}
       this.agent = agent
       if (
         this.agent &&
@@ -48,7 +52,8 @@ define([
         } sourcetype=wazuh rule.gdpr{}="*"| stats count by "rule.gdpr{}" | spath "rule.gdpr{}" | fields - count`,
         'rule.gdpr{}',
         '$form.gdpr$',
-        'dropDownInput'
+        'dropDownInput',
+        this.scope
       )
       this.getFilters = this.currentDataService.getSerializedFilters
       this.dropdownInstance = this.dropdown.getElement()
@@ -77,38 +82,107 @@ define([
           `${
             this.filters
           } sourcetype=wazuh rule.gdpr{}="$gdpr$"  | stats count by rule.gdpr{}`,
-          'gdprRequirementsVizz'
+          'gdprRequirementsVizz',
+          this.scope
         ),
         new PieChart(
           'groupsVizz',
           `${
             this.filters
           } sourcetype=wazuh rule.gdpr{}="$gdpr$" | stats count by rule.groups`,
-          'groupsVizz'
+          'groupsVizz',
+          this.scope
         ),
         new PieChart(
           'agentsVizz',
           `${
             this.filters
           } sourcetype=wazuh rule.gdpr{}="$gdpr$" | stats count by agent.name`,
-          'agentsVizz'
+          'agentsVizz',
+          this.scope
         ),
         new ColumnChart(
           'requirementsByAgentVizz',
           `${
             this.filters
           } sourcetype=wazuh rule.gdpr{}="$gdpr$" agent.name=*| chart  count(rule.gdpr{}) by rule.gdpr{},agent.name`,
-          'requirementsByAgentVizz'
+          'requirementsByAgentVizz',
+          this.scope
         ),
         new Table(
           'alertsSummaryVizz',
           `${
             this.filters
           } sourcetype=wazuh rule.gdpr{}="$gdpr$" | stats count sparkline by agent.name, rule.gdpr{}, rule.description | sort count DESC | rename agent.name as "Agent Name", rule.gdpr{} as Requirement, rule.description as "Rule description", count as Count`,
-          'alertsSummaryVizz'
+          'alertsSummaryVizz',
+          this.scope
         )
       ]
 
+      this.alertsSummaryTable = new rawTableDataService(
+        'alertsSummaryTable',
+        `${
+          this.filters
+        } sourcetype=wazuh rule.gdpr{}="$gdpr$" | stats count sparkline by agent.name, rule.gdpr{}, rule.description | sort count DESC | rename agent.name as "Agent Name", rule.gdpr{} as Requirement, rule.description as "Rule description", count as Count`,
+        'alertsSummaryTableToken',
+        '$result$',
+        this.scope
+      )
+      this.vizz.push(this.alertsSummaryTable)
+
+      this.alertsSummaryTable.getSearch().on('result', (result) => {
+        this.tableResults['Alerts Summary'] = result
+      })
+     
+      // Set agent info
+      try {
+        this.agentReportData = {
+          ID: this.agent.data.data.id,
+          Name: this.agent.data.data.name,
+          IP: this.agent.data.data.ip,
+          Version: this.agent.data.data.version,
+          Manager: this.agent.data.data.manager,
+          OS: this.agent.data.data.os.name,
+          dateAdd: this.agent.data.data.dateAdd,
+          lastKeepAlive: this.agent.data.data.lastKeepAlive,
+          group: this.agent.data.data.group.toString()
+        }
+      } catch (error) {
+        this.agentReportData = false
+      }
+
+      /**
+       * Generates report
+       */
+      this.scope.startVis2Png = () =>
+      this.reportingService.startVis2Png('agents-gdpr', 'GDPR', this.filters, [
+        'gdprRequirementsVizz',
+        'groupsVizz',
+        'agentsVizz',
+        'requirementsByAgentVizz',
+        'alertsSummaryVizz'
+      ],
+      {},//Metrics,
+      this.tableResults,
+      this.agentReportData)
+      
+
+
+      this.scope.$on('loadingReporting', (event, data) => {
+        this.scope.loadingReporting = data.status
+      })
+
+      this.scope.$on("checkReportingStatus", () => {
+        this.vizzReady = !this.vizz.filter( v => {
+          return v.finish === false
+        }).length
+        if (this.vizzReady) { 
+          this.scope.loadingVizz = false
+        } else { 
+          this.scope.loadingVizz = true
+        }
+        if (!this.scope.$$phase) this.scope.$digest()
+      })
       /**
        * When controller is destroyed
        */