+++ SplunkAppForWazuh/appserver/static/js/controllers/agents/scap/agentsOpenScapCtrl.js
@@ -18,7 +18,8 @@ define([
   '../../../services/visualizations/table/table',
   '../../../services/visualizations/inputs/time-picker',
   '../../../services/visualizations/inputs/dropdown-input',
-  '../../../services/visualizations/search/search-handler'
+  '../../../services/visualizations/search/search-handler',
+  '../../../services/rawTableData/rawTableDataService'
 ], function(
   app,
   PieChart,
@@ -27,7 +28,8 @@ define([
   Table,
   TimePicker,
   Dropdown,
-  SearchHandler
+  SearchHandler,
+  rawTableDataService
 ) {
   'use strict'
 
@@ -38,13 +40,16 @@ define([
      * @param {Object} $scope
      * @param {Object} $currentDataService
      * @param {Object} $state
+     * @param {*} $reportingService 
      * @param {Object} agent
      */
 
-    constructor($urlTokenModel, $scope, $currentDataService, $state, agent) {
+    constructor($urlTokenModel, $scope, $currentDataService, $state, agent, $reportingService) {
       this.urlTokenModel = $urlTokenModel
       this.scope = $scope
       this.currentDataService = $currentDataService
+      this.reportingService = $reportingService
+      this.tableResults = {}
       this.state = $state
       this.agent = agent
       this.currentDataService.addFilter(
@@ -71,7 +76,8 @@ define([
         } sourcetype=wazuh  rule.groups!="syslog" oscap.scan.profile.title=* | stats count by oscap.scan.profile.title | sort oscap.scan.profile.title ASC|fields - count`,
         'oscap.scan.profile.title',
         '$form.profile$',
-        'dropDownInput'
+        'dropDownInput',
+        this.scope
       )
       this.dropdownInstance = this.dropdown.getElement()
       this.submittedTokenModel = this.urlTokenModel.getSubmittedTokenModel()
@@ -127,59 +133,135 @@ define([
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups!="syslog" oscap.scan.profile.title="$profile$" | top agent.name`,
-          'agentsVizz'
+          'agentsVizz',
+          this.scope
         ),
         new PieChart(
           'profilesVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups!="syslog" oscap.scan.profile.title="$profile$" | top oscap.scan.profile.title`,
-          'profilesVizz'
+          'profilesVizz',
+          this.scope
         ),
         new BarChart(
           'contentVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups!="syslog" oscap.scan.profile.title="$profile$" | top oscap.scan.content`,
-          'contentVizz'
+          'contentVizz',
+          this.scope
         ),
         new PieChart(
           'severityVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups!="syslog" oscap.scan.profile.title="$profile$" | top oscap.check.severity`,
-          'severityVizz'
+          'severityVizz',
+          this.scope
         ),
         new AreaChart(
           'top5AgentsSHVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.scan.profile.title="$profile$" oscap.check.severity="high" | chart count by agent.name`,
-          'top5AgentsSHVizz'
+          'top5AgentsSHVizz',
+          this.scope
         ),
         new PieChart(
           'top10AleertsVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups="oscap-result" oscap.scan.profile.title="$profile$" | top oscap.check.title`,
-          'top10AleertsVizz'
+          'top10AleertsVizz',
+          this.scope
         ),
         new PieChart(
           'top10HRAlertsVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" rule.groups="oscap-result"  oscap.check.severity="high" oscap.scan.profile.title="$profile$" | top oscap.check.title`,
-          'top10HRAlertsVizz'
+          'top10HRAlertsVizz',
+          this.scope
         ),
         new Table(
           'alertsSummaryVizz',
           `${
             this.filters
           } sourcetype=wazuh oscap.check.result="fail" oscap.scan.profile.title="$profile$" | stats count by agent.name, oscap.check.title, oscap.scan.profile.title, oscap.scan.id, oscap.scan.content | sort count DESC | rename agent.name as "Agent name", oscap.check.title as Title, oscap.scan.profile.title as Profile, oscap.scan.id as "Scan ID", oscap.scan.content as Content`,
-          'alertsSummaryVizz'
+          'alertsSummaryVizz',
+          this.scope
         )
       ]
 
+      this.alertsSummaryTable = new rawTableDataService(
+        'alertsSummaryTable',
+        `${
+          this.filters
+        } sourcetype=wazuh oscap.check.result="fail" oscap.scan.profile.title="$profile$" | stats count by agent.name, oscap.check.title, oscap.scan.profile.title, oscap.scan.id, oscap.scan.content | sort count DESC | rename agent.name as "Agent name", oscap.check.title as Title, oscap.scan.profile.title as Profile, oscap.scan.id as "Scan ID", oscap.scan.content as Content`,
+        'alertsSummaryTableToken',
+        '$result$',
+        this.scope
+      )
+      this.vizz.push(this.alertsSummaryTable)
+
+      this.alertsSummaryTable.getSearch().on('result', (result) => {
+        this.tableResults['Alerts Summary'] = result
+      })
+
+      // Set agent info
+      try {
+        this.agentReportData = {
+          ID: this.agent.data.data.id,
+          Name: this.agent.data.data.name,
+          IP: this.agent.data.data.ip,
+          Version: this.agent.data.data.version,
+          Manager: this.agent.data.data.manager,
+          OS: this.agent.data.data.os.name,
+          dateAdd: this.agent.data.data.dateAdd,
+          lastKeepAlive: this.agent.data.data.lastKeepAlive,
+          group: this.agent.data.data.group.toString()
+        }
+      } catch (error) {
+        this.agentReportData = false
+      }
+
+      /**
+       * Generates report
+       */
+      this.scope.startVis2Png = () =>
+      this.reportingService.startVis2Png('agents-openscap', 'Open SCAP', this.filters, [
+        'agentsVizz',
+        'profilesVizz',
+        'contentVizz',
+        'severityVizz',
+        'top5AgentsSHVizz',
+        'top10AleertsVizz',
+        'top10HRAlertsVizz',
+        'alertsSummaryVizz'
+      ],
+      this.reportMetrics,
+      this.tableResults,
+      this.agentReportData
+      )
+
+      this.scope.$on('loadingReporting', (event, data) => {
+        this.scope.loadingReporting = data.status
+      })
+
+      this.scope.$on("checkReportingStatus", () => {
+        this.vizzReady = !this.vizz.filter( v => {
+          return v.finish === false
+        }).length
+        if (this.vizzReady) { 
+          this.scope.loadingVizz = false
+          this.setReportMetrics()
+        } else { 
+          this.scope.loadingVizz = true
+        }
+        if (!this.scope.$$phase) this.scope.$digest()
+      })
+      
       /**
        * When controller is destroyed
        */
@@ -220,6 +302,17 @@ define([
      */
     getAgentStatusClass(agentStatus) {
       agentStatus === 'Active' ? 'teal' : 'red'
+    }
+
+    /**
+     * Set report metrics
+     */
+    setReportMetrics() {
+      this.reportMetrics = {
+        'Last score': this.scope.scapLastScore,
+        'Highest score': this.scope.scapHighestScore,
+        'Lowest score': this.scope.scapLowestScore
+      }
     }
 
     /**