+++ SplunkAppForWazuh/appserver/static/js/controllers/overview/general/overviewGeneralCtrl.js
@@ -6,7 +6,8 @@ define([
   '../../../services/visualizations/chart/area-chart',
   '../../../services/visualizations/table/table',
   '../../../services/visualizations/inputs/time-picker',
-  '../../../services/visualizations/search/search-handler'
+  '../../../services/visualizations/search/search-handler',
+  '../../../services/rawTableData/rawTableDataService',
 ], function(
   app,
   LinearChart,
@@ -15,7 +16,8 @@ define([
   AreaChart,
   Table,
   TimePicker,
-  SearchHandler
+  SearchHandler,
+  rawTableDataService
 ) {
   'use strict'
 
@@ -30,6 +32,7 @@ define([
      * @param {*} $requestService 
      * @param {Object} pollingState 
      * @param {*} $reportingService 
+     * @param {*} $rootScope 
      */
     constructor(
       $urlTokenModel,
@@ -39,13 +42,16 @@ define([
       $notificationService,
       $requestService,
       pollingState,
-      $reportingService
+      $reportingService,
+      $rootScope
     ) {
       this.currentDataService = $currentDataService
+      this.rootScope = $rootScope
       this.filters = this.currentDataService.getSerializedFilters()
       this.scope = $scope
       this.reportingService = $reportingService
       this.apiReq = $requestService.apiReq
+      this.tableResults = {}
       this.timePicker = new TimePicker(
         '#timePicker',
         $urlTokenModel.handleValueChange
@@ -120,33 +126,53 @@ define([
           `${
             this.filters
           } sourcetype=wazuh rule.level=*| timechart count by rule.level`,
-          'alertLevEvoVizz'
+          'alertLevEvoVizz',
+          this.scope
         ),
         new ColumnChart(
           'alertsVizz',
           `${this.filters} sourcetype=wazuh | timechart span=2h count`,
-          'alertsVizz'
+          'alertsVizz',
+          this.scope
         ),
         new PieChart(
           'top5AgentsVizz',
           `${this.filters} sourcetype=wazuh | top agent.name`,
-          'top5AgentsVizz'
+          'top5AgentsVizz',
+          this.scope
         ),
         new AreaChart(
           'alertsEvoTop5Agents',
           `${
             this.filters
           } sourcetype=wazuh | timechart span=1h limit=5 useother=f count by agent.name`,
-          'alertsEvoTop5Agents'
+          'alertsEvoTop5Agents',
+          this.scope
         ),
         new Table(
           'agentsSummaryVizz',
           `${
             this.filters
           } sourcetype=wazuh |stats count sparkline by rule.id, rule.description, rule.level | sort rule.level DESC | rename rule.id as "Rule ID", rule.description as "Description", rule.level as Level, count as Count`,
-          'agentsSummaryVizz'
+          'agentsSummaryVizz',
+          this.scope
         )
       ]
+
+      this.agentsSummaryTable = new rawTableDataService(
+        'agentsSummaryTable',
+        `${
+          this.filters
+        } sourcetype=wazuh |stats count sparkline by rule.id, rule.description, rule.level | sort rule.level DESC | rename rule.id as "Rule ID", rule.description as "Description", rule.level as Level, count as Count`,
+        'agentsSummaryTableToken',
+        '$result$',
+        this.scope
+      )
+      this.vizz.push(this.agentsSummaryTable)
+
+      this.agentsSummaryTable.getSearch().on('result', (result) => {
+        this.tableResults['Agents Summary'] = result
+      })
     }
 
     /**
@@ -197,19 +223,24 @@ define([
             `${this.agentsStatusFilter} status=* | timechart span=${
               this.spanTime
             } cont=FALSE count by status usenull=f`,
-            `agentStatus`
+            `agentStatus`,
+            this.scope
           )
         )
       }
 
       this.scope.startVis2Png = () =>
-        this.reportingService.startVis2Png('overview-general', [
+        this.reportingService.startVis2Png('overview-general', 'Security events', this.filters, 
+        [
           'alertLevEvoVizz',
           'alertsVizz',
           'top5AgentsVizz',
           'alertsEvoTop5Agents',
           'agentsSummaryVizz'
-        ])
+        ],
+        this.reportMetrics,
+        this.tableResults
+        )
 
       this.scope.$on('$destroy', () => {
         this.timePicker.destroy()
@@ -219,6 +250,19 @@ define([
       this.scope.$on('loadingReporting', (event, data) => {
         this.scope.loadingReporting = data.status
       })
+
+      this.scope.$on("checkReportingStatus", () => {
+          this.vizzReady = !this.vizz.filter( v => {
+            return v.finish === false
+          }).length
+          if (this.vizzReady) { 
+            this.scope.loadingVizz = false
+            this.setReportMetrics()
+          } else { 
+            this.scope.loadingVizz = true
+          }
+          if (!this.scope.$$phase) this.scope.$digest()
+      })
     }
 
     /**
@@ -227,6 +271,18 @@ define([
     launchSearches() {
       this.filters = this.currentDataService.getSerializedFilters()
       this.state.reload()
+    }
+
+    /**
+     * Set report metrics
+     */
+    setReportMetrics() {
+      this.reportMetrics = {
+        'Alerts': this.scope.totalAlerts,
+        'Level 12 or above alerts': this.scope.levelTwelve,
+        'Authentication failure': this.scope.authFailure,
+        'Authentication success': this.scope.authSuccess
+      }
     }
   }