+++ SplunkAppForWazuh/appserver/static/js/controllers/overview/policy-monitoring/overviewPolicyMonitoringCtrl.js
@@ -3,8 +3,9 @@ define([
   '../../../services/visualizations/chart/pie-chart',
   '../../../services/visualizations/chart/area-chart',
   '../../../services/visualizations/table/table',
-  '../../../services/visualizations/inputs/time-picker'
-], function(app, PieChart, AreaChart, Table, TimePicker) {
+  '../../../services/visualizations/inputs/time-picker',
+  '../../../services/rawTableData/rawTableDataService'
+], function(app, PieChart, AreaChart, Table, TimePicker, rawTableDataService) {
   'use strict'
   class PM {
     /**
@@ -13,11 +14,14 @@ define([
      * @param {*} $scope 
      * @param {*} $currentDataService 
      * @param {*} $state 
+     * @param {*} $reportingService
      */
-    constructor($urlTokenModel, $scope, $currentDataService, $state) {
+    constructor($urlTokenModel, $scope, $currentDataService, $state, $reportingService) {
       this.scope = $scope
       this.urlTokenModel = $urlTokenModel
       this.state = $state
+      this.reportingService = $reportingService
+      this.tableResults = {}
       this.getFilters = $currentDataService.getSerializedFilters
       this.filters = this.getFilters()
       this.timePicker = new TimePicker(
@@ -33,37 +37,57 @@ define([
           `${
             this.filters
           } sourcetype=wazuh "rule.groups"="rootcheck" rule.description=* | timechart span=1h count by rule.description`,
-          'elementOverTime'
+          'elementOverTime',
+          this.scope
         ),
         new PieChart(
           'cisRequirements',
           `${
             this.filters
           } sourcetype=wazuh "rule.groups"="rootcheck" rule.cis{}=* | top  rule.cis{}`,
-          'cisRequirements'
+          'cisRequirements',
+          this.scope
         ),
         new PieChart(
           'topPciDss',
           `${
             this.filters
           } sourcetype=wazuh "rule.groups"="rootcheck" rule.pci_dss{}=* | top  rule.pci_dss{}`,
-          'topPciDss'
+          'topPciDss',
+          this.scope
         ),
         new AreaChart(
           'eventsPerAgent',
           `${
             this.filters
           } sourcetype=wazuh "rule.groups"="rootcheck" | timechart span=2h count by agent.name`,
-          'eventsPerAgent'
+          'eventsPerAgent',
+          this.scope
         ),
         new Table(
           'alertsSummary',
           `${
             this.filters
           } sourcetype=wazuh "rule.groups"="rootcheck" |stats count sparkline by agent.name, rule.description, title | sort count DESC | rename rule.description as "Rule description", agent.name as Agent, title as Control`,
-          'alertsSummary'
+          'alertsSummary',
+          this.scope
         )
       ]
+
+      this.alertsSummaryTable = new rawTableDataService(
+        'alertsSummaryTable',
+        `${
+          this.filters
+        } sourcetype=wazuh "rule.groups"="rootcheck" |stats count sparkline by agent.name, rule.description, title | sort count DESC | rename rule.description as "Rule description", agent.name as Agent, title as Control`,
+        'alertsSummaryTableToken',
+        '$result$',
+        this.scope
+      )
+      this.vizz.push(this.alertsSummaryTable)
+
+      this.alertsSummaryTable.getSearch().on('result', (result) => {
+        this.tableResults['Alerts Summary'] = result
+      })
 
       this.scope.$on('deletedFilter', () => {
         this.launchSearches()
@@ -79,6 +103,37 @@ define([
         this.timePicker.destroy()
         this.vizz.map(vizz => vizz.destroy())
       })
+
+      /**
+       * Generates report
+       */
+      this.scope.startVis2Png = () =>
+      this.reportingService.startVis2Png('overview-pm', 'Policity monitoring', this.filters, [
+        'elementOverTime',
+        'cisRequirements',
+        'topPciDss',
+        'eventsPerAgent',
+        'alertsSummary'
+      ],
+      {},//Metrics
+      this.tableResults)
+
+      this.scope.$on('loadingReporting', (event, data) => {
+        this.scope.loadingReporting = data.status
+      })
+
+      this.scope.$on("checkReportingStatus", () => {
+        this.vizzReady = !this.vizz.filter( v => {
+          return v.finish === false
+        }).length
+        if (this.vizzReady) { 
+          this.scope.loadingVizz = false
+        } else { 
+          this.scope.loadingVizz = true
+        }
+        if (!this.scope.$$phase) this.scope.$digest()
+    })
+
     }
 
     /**