+++ SplunkAppForWazuh/appserver/static/js/services/visualizations/chart/linear-chart.js
@@ -10,8 +10,9 @@ define(['splunkjs/mvc/simplexml/element/chart', '../viz/viz'], function(
      * @param {String} id
      * @param {String} search
      * @param {String} attachedElement
+     * @param {scope} scope
      */
-    constructor(id, search, attachedElement) {
+    constructor(id, search, attachedElement, scope) {
       super(
         new ChartElement(
           {
@@ -25,7 +26,8 @@ define(['splunkjs/mvc/simplexml/element/chart', '../viz/viz'], function(
           { tokens: true, tokenNamespace: 'submitted' }
         ).render(),
         id,
-        search
+        search,
+        scope
       )
     }
   }